Moving MergeQueue to index package as it now has shared utility
