+++ spark/src/main/scala/geotrellis/spark/io/index/zcurve/Z2.scala
@@ -1,5 +1,7 @@
 package geotrellis.spark.io.index.zcurve
 
+import geotrellis.spark.io.index.MergeQueue
+
 class Z2(val z: Long) extends AnyVal {
   import Z2._
